feat: Slack-style subdomain workspaces + independent domain
Major architectural changes:
- Subdomain routing: <workspace>.scitex-orochi.com (middleware-based)
- Bare domain landing page with create/find workspace
- Split views.py into views/ package (auth, workspace, api, landing)
- WebSocket reads workspace from Host header instead of URL path
- SSO visitor filtering: reject non-member scitex.ai accounts
- Cloudflare wildcard DNS + tunnel for *.scitex-orochi.com
- Session cookies shared across subdomains
- Old /workspace/<slug>/ URLs redirect to subdomains
