+++ hub/static/hub/simple.css
@@ -21,9 +21,78 @@ body {
   color: #fff;
   margin-bottom: 0.5rem;
 }
+.logo {
+  width: 64px;
+  height: 64px;
+  margin-bottom: 0.5rem;
+  border-radius: 12px;
+}
 .card p {
   color: #888;
 }
 a {
   color: #4a9eff;
+}
+.subtitle {
+  color: #888;
+  margin-top: 0;
+  margin-bottom: 1.5rem;
+}
+.btn {
+  padding: 10px 16px;
+  border-radius: 4px;
+  border: 1px solid #555;
+  background: #333;
+  color: #d1d2d3;
+  cursor: pointer;
+  text-decoration: none;
+  font-size: 0.9rem;
+}
+.btn-primary {
+  background: #4a9eff;
+  border-color: #4a9eff;
+  color: #fff;
+}
+.btn-primary:hover {
+  background: #3a8eef;
+}
+.btn-block {
+  display: block;
+  text-align: center;
+  margin-bottom: 10px;
+}
+.flex-row {
+  display: flex;
+  gap: 8px;
+  margin-top: 6px;
+}
+.flex-1 {
+  flex: 1;
+}
+.mt-sm {
+  margin-top: 12px;
+}
+.msg-error {
+  color: #e74c3c;
+}
+.msg-success {
+  color: #2ecc71;
+}
+label {
+  display: block;
+  text-align: left;
+  color: #888;
+  font-size: 0.85rem;
+}
+input[type="text"],
+input[type="email"],
+input[type="password"] {
+  background: #2a2d31;
+  border: 1px solid #444;
+  border-radius: 4px;
+  padding: 10px;
+  color: #d1d2d3;
+  font-size: 0.9rem;
+  width: 100%;
+  box-sizing: border-box;
 }
