+++ hub/templates/hub/sso_rejected.html
@@ -1,0 +1,23 @@
+{% load static %}
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <title>SciTeX Orochi - Sign In Not Allowed</title>
+        <link rel="stylesheet" href="{% static 'hub/simple.css' %}">
+    </head>
+    <body>
+        <div class="card">
+            <h1>Sign In Not Allowed</h1>
+            <p>
+                Your SciTeX account (<strong>{{ username }}</strong>) is a
+                <strong>{{ user_type }}</strong> account, which cannot sign in
+                to Orochi.
+            </p>
+            <p>Only registered SciTeX members can use SciTeX SSO.</p>
+            <p>
+                You can still <a href="{% url 'signin' %}">sign up</a> for a local Orochi account.
+            </p>
+        </div>
+    </body>
+</html>
