+++ hub/views/__init__.py
@@ -1,0 +1,26 @@
+"""Hub views package — re-exports all view functions."""
+
+from hub.views.api import (  # noqa: F401
+    api_channels,
+    api_history,
+    api_messages,
+    api_stats,
+    api_workspaces,
+)
+from hub.views.auth import (  # noqa: F401
+    accept_invite_view,
+    create_workspace_view,
+    index,
+    signin_view,
+    signout_view,
+    signup_view,
+)
+from hub.views.landing import (  # noqa: F401
+    find_workspace_view,
+    landing_page,
+    redirect_old_workspace_url,
+)
+from hub.views.workspace import (  # noqa: F401
+    workspace_dashboard,
+    workspace_settings_view,
+)
